#19 opening/StudyPeriod 컴포넌트 분리
